chore: move ClientConfig construction to tls/client_config.rs
